Merge: menu 인증/인가 설정 및 upstream 병합

+++ src/menu/menu.service.ts
@@ -1,10 +1,12 @@
 import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
-import { Menu, Prisma } from '@prisma/client';
+import { Menu, Owner, Prisma } from '@prisma/client';
 import { PrismaService } from 'src/prisma/prisma.service';
+import { StoreService } from 'src/store/store.service';
 import { CreateMenuDto } from './dto/create-menu.dto';
 import { UpdateMenuDto } from './dto/update-menu.dto';
-import { StoreService } from 'src/store/store.service';
 
+//Todo: 중복되는 코드 정리 필요
+//! 윤주님이 storeService 메소드 명 수정시 `findOneComment`도 같이 수정 해야함
 @Injectable()
 export class MenuService {
   constructor(
@@ -13,11 +15,17 @@ export class MenuService {
   ) {}
 
   //* 메뉴 생성
-  async createMenu(createMenuDto: CreateMenuDto): Promise<Menu> {
+  async createMenu(createMenuDto: CreateMenuDto, user: Owner): Promise<Menu> {
+
     // 업장 확인
     const store = await this.storeService.findOneComment(createMenuDto.StoreId);
     if (!store) {
       throw new HttpException('업장 정보가 존재하지 않습니다.', HttpStatus.NOT_FOUND);
+    }
+    
+    // 접근 권한 확인
+    if(store.OwnerId !== user.id){
+      throw new HttpException('접근 권한이 없습니다.', HttpStatus.FORBIDDEN);
     }
 
     // 메뉴 중복 확인
@@ -74,9 +82,27 @@ export class MenuService {
   }
 
   //* 메뉴 수정
-  async updateMenu(updateMenuDto: UpdateMenuDto): Promise<Menu> {
-    // 가게 및 메뉴 확인
-    await this.getMenu({ id: updateMenuDto.menuId, StoreId: updateMenuDto.StoreId });
+  async updateMenu(updateMenuDto: UpdateMenuDto, user: Owner): Promise<Menu> {
+    // 업장 확인
+    const store = await this.storeService.findOneComment(Number(updateMenuDto.StoreId));
+    if (!store) {
+      throw new HttpException('업장 정보가 존재하지 않습니다.', HttpStatus.NOT_FOUND);
+    }
+
+    // 접근 권한 확인
+    if(store.OwnerId !== user.id){
+        throw new HttpException('접근 권한이 없습니다.', HttpStatus.FORBIDDEN);
+    }
+  
+    // 메뉴 확인
+    const isMenu = await this.prisma.menu.findUnique({
+      where: {
+        id:updateMenuDto.menuId,
+      }
+    });
+    if(!isMenu){
+      throw new HttpException('메뉴가 존재하지 않습니다.', HttpStatus.NOT_FOUND);
+    }
 
     // 메뉴 중복 확인
     const menu = await this.prisma.menu.findFirst({
@@ -87,7 +113,7 @@ export class MenuService {
     if (menu) {
       throw new HttpException('이미 존재하는 메뉴입니다.', HttpStatus.BAD_REQUEST);
     }
-
+    
     // 메뉴 수정
     return this.prisma.menu.update({
       where: {
@@ -103,9 +129,27 @@ export class MenuService {
   }
 
   //* 메뉴 삭제
-  async deleteMenu(where: Prisma.MenuWhereUniqueInput): Promise<Menu> {
-    // 가게 및 메뉴 확인
-    await this.getMenu({ id: where.id, StoreId: where.StoreId });
+  async deleteMenu(where: Prisma.MenuWhereUniqueInput, user: Owner): Promise<Menu> {
+    // 업장 확인
+    const store = await this.storeService.findOneComment(Number(where.StoreId));
+    if (!store) {
+      throw new HttpException('업장 정보가 존재하지 않습니다.', HttpStatus.NOT_FOUND);
+    }
+
+    // 접근 권한 확인
+    if(store.OwnerId !== user.id){
+        throw new HttpException('접근 권한이 없습니다.', HttpStatus.FORBIDDEN);
+    }
+  
+    // 메뉴 확인
+    const isMenu = await this.prisma.menu.findUnique({
+      where: {
+        id: where.id,
+      }
+    });
+    if(!isMenu){
+      throw new HttpException('메뉴가 존재하지 않습니다.', HttpStatus.NOT_FOUND);
+    }
 
     // 메뉴 삭제
     return this.prisma.menu.delete({
